Update main
Using a switch statement instead of multiple if/else if statements: Since there are multiple commands to check, it might be more efficient and cleaner to use a switch statement instead of a long chain of if/else if statements.

Checking the return value of the system function: The system function returns an integer value that represents the exit status of the executed command. You should check this value to make sure the command was executed successfully.

Adding error handling for the delete command: The delete command should have error handling for cases where the file does not exist or the user does not have permission to delete the file.